+++ src/Exceptions/EmailAddressInvalidArgumentException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * This file is a part of the datatypes package.
+ *
+ * Read more at https://phpdatatypes.com/
+ */
+
+namespace DataTypes\Exceptions;
+
+/**
+ * Exception used when passed invalid argument in EmailAddress class.
+ *
+ * @since 1.1.0
+ */
+class EmailAddressInvalidArgumentException extends \InvalidArgumentException
+{
+}