+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Added
+- EmailAddress class.
 
 ## [1.0.1] - 2015-05-29
 - Fixed [#4](https://github.com/themichaelhall/datatypes/issues/4) - Url with name and/or password should not be treated as invalid.